The code that sets the CTCT_SDK_VERSION in Java should be moved to a class called SdkVersion.
